requirements.txt
asgiref==3.5.2
astroid==2.11.7
attrs==22.1.0
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
coverage==6.4.3
dill==0.3.5.1
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
fire==0.4.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pylint==2.14.5
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
tomli==2.0.1
tomlkit==0.11.3
typing_extensions==4.3.0
urllib3==1.26.11
wrapt==1.14.1
